backend/requirements.txt
absl-py==2.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Cython==3.2.4
easydict==1.13
fastapi==0.135.1
flatbuffers==25.12.19
fonttools==4.62.0
h11==0.16.0
idna==3.11
ImageIO==2.37.3
insightface==0.7.3
joblib==1.5.3
kiwisolver==1.5.0
lazy-loader==0.5
matplotlib==3.10.8
mediapipe==0.10.33
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
nvidia-cublas-cu12==12.9.1.4
nvidia-cudnn-cu12==9.20.0.48
onnx==1.20.1
onnxruntime-gpu==1.24.4
opencv-contrib-python==4.13.0.92
opencv-python-headless==4.13.0.92
outcome==1.3.0.post0
packaging==26.0
pillow==12.1.1
prettytable==3.17.0
protobuf==7.34.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
selenium==4.41.0
simsimd==6.5.16
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.5
starlette==0.52.1
stringzilla==4.6.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2026.3.3
tqdm==4.67.3
trio==0.33.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
wcwidth==0.6.0
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
